requirements.txt
alembic==1.6.5
APScheduler==3.7.0
certifi==2021.5.30
greenlet==1.1.0
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
tornado==6.1
tzlocal==2.1
